backend/requirements.txt
django-cors-headers==3.7.0
django-paypal==1.1.1
djangorestframework==3.12.4
pip-chill==1.0.1
